Organize project folder - Add command folder - Add listener folder - Move commands and listeners to folders - Move AnnouncementManager.java
