+++ toolkitplugin/src/main/java/org/fourz/rvnktools/RVNKTools.java
@@ -5,27 +5,30 @@ import java.util.logging.Level;
 import org.bukkit.Bukkit;
 import org.bukkit.plugin.java.JavaPlugin;
 import org.checkerframework.checker.units.qual.A;
-import org.fourz.rvnktools.DiscordCommand;
-import org.fourz.rvnktools.listeners.MickyHatPlaceListener;
-import org.fourz.rvnktools.AnnouncementManager;
+import org.fourz.rvnktools.announcementManager.AnnouncementManager;
+import org.fourz.rvnktools.command.DiscordCommand;
+import org.fourz.rvnktools.command.EventsCommand;
+import org.fourz.rvnktools.command.PingCommand;
+import org.fourz.rvnktools.listener.JoinListener;
+import org.fourz.rvnktools.listener.MickyHatPlaceListener;
 
 public class RVNKTools extends JavaPlugin {
 
-    private AnnouncementManager announcementManager;    
+    private AnnouncementManager announcementManager;
 
     @Override
     public void onEnable() {
-		
+
         // Initialize AnnouncementManager
         announcementManager = new AnnouncementManager(this);
 
         // Code that runs when the plugin is enabled
-		getServer().getPluginManager().registerEvents(new JoinListener(), this);
+        getServer().getPluginManager().registerEvents(new JoinListener(), this);
         getServer().getPluginManager().registerEvents(new MickyHatPlaceListener(), this);
         getLogger().info("yoo");
-        
-		// Register commands
-		this.getCommand("ping").setExecutor(new PingCommand());
+
+        // Register commands
+        this.getCommand("ping").setExecutor(new PingCommand());
         this.getCommand("tps").setExecutor(this);
         this.getCommand("events").setExecutor(new EventsCommand());
         this.getCommand("discord").setExecutor(new DiscordCommand());
@@ -39,5 +42,3 @@ public class RVNKTools extends JavaPlugin {
         getLogger().info("RVNK Toolkit has been disabled.");
     }
 }
-
-
