+++ toolkitplugin/src/main/java/org/fourz/rvnktools/command/EventsCommand.java
@@ -1,4 +1,4 @@
-package org.fourz.rvnktools;
+package org.fourz.rvnktools.command;
 
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
