+++ toolkitplugin/src/main/java/org/fourz/rvnktools/listener/JoinListener.java
@@ -1,13 +1,12 @@
-package org.fourz.rvnktools;
+package org.fourz.rvnktools.listener;
 
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.entity.Player;
 import org.bukkit.event.player.PlayerJoinEvent;
 
-
 public class JoinListener implements Listener {
-	
+
     @EventHandler
     public void onPlayerJoin(PlayerJoinEvent event) {
         Player player = event.getPlayer();
