+++ toolkitplugin/src/main/java/org/fourz/rvnktools/listener/MickyHatPlaceListener.java
@@ -1,4 +1,4 @@
-package org.fourz.rvnktools.listeners;
+package org.fourz.rvnktools.listener;
 
 import org.bukkit.Material;
 import org.bukkit.event.EventHandler;
@@ -15,7 +15,7 @@ public class MickyHatPlaceListener implements Listener {
         // Ensure it's a block placement action
         if (event.getAction() == Action.RIGHT_CLICK_BLOCK) {
             ItemStack item = event.getItem();
-            
+
             if (item != null && item.getType() == Material.CARVED_PUMPKIN) {
                 ItemMeta meta = item.getItemMeta();
                 if (meta != null && meta.hasCustomModelData()) {
